% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_niche.R
\name{model_path}
\alias{model_path}
\title{Get path to a Stan model file}
\usage{
model_path(
  model = c("presence_only", "bm_constant", "bm_evolving", "ou_constant", "ou_evolving",
    "null")
)
}
\arguments{
\item{model}{Character string specifying the model. Currently only
\code{"presence_only"} is supported in arfun.}
}
\value{
A character string with the full path to the Stan model file.
}
\description{
Returns the full path to a Stan model file included in the package.
}
\examples{
model_path("presence_only")

}
